rename preact utils o optional signal and move it
