+++ app/ts/utils/OptionalSignal.tsx
